+++ app/src/main/java/com/undistract/ui/profile/ProfileFormScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import com.undistract.AppSelectionDialog
-import com.undistract.ProfileFormViewModel
+import com.undistract.ui.profile.ProfileFormViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
